Abacus v1.0.0 Abacus.SystemMetric

Functions

unit(name) (macro)

A macro to generate the base of the system. This unit is the reference of each other units.

For example :

defmodule Example do 
  use Abacus.SystemMetric

  unit :cm
end

unit(name, expr) (macro)

A macro to generate an unit using the base as a reference. This is referenced by a name (:km for example) and by a reference to the base, in the case of :km in a system referenced by :cm : 100000.

For example:

defmodule Example do 
  use Abacus.SystemMetric

  unit :cm
  unit :m, 100 # (100 cm  == 1 m)
  unit :dm, 10 # (10 cm == 1 dm)
end